Move Substitutor and AlphaConverting_Substitutor to separate files
