+++ Source/Dafny/DafnyAst.cs
@@ -9639,7 +9639,7 @@ namespace Microsoft.Dafny {
         substMap.Add(oldVars[i], id);
       }
 
-      Translator.Substituter sub = new Translator.Substituter(null, substMap, typeMap);
+      Substituter sub = new Substituter(null, substMap, typeMap);
       return sub.Substitute(e);
     }
