+++ Source/Dafny/Translators/AlphaConverting_Substituter.cs
@@ -1,0 +1,35 @@
+using System.Collections.Generic;
+using System.Diagnostics.Contracts;
+
+namespace Microsoft.Dafny
+{
+  /// <summary>
+  /// This substituter performs substitutions in such a way that it's okay to print the resulting expression without a human getting confused.
+  /// More precisely, bound variables first gets alpha-renamed.  Also, "this" is never left implicit, including in the
+  /// case where "receiverReplacement" is given as ImplicitThisExpr (but no attempt is made to substitute for all ImplicitThisExpr's in
+  /// "receiverReplacement" and the range of "substMap").
+  /// </summary>
+  public class AlphaConverting_Substituter : Substituter
+  {
+    ISet<string> namesToAvoid = new HashSet<string>();
+    public AlphaConverting_Substituter(Expression receiverReplacement, Dictionary<IVariable, Expression> substMap, Dictionary<TypeParameter, Type> typeMap)
+      : base(receiverReplacement is ImplicitThisExpr ? new ThisExpr(receiverReplacement.tok) { Type = receiverReplacement.Type } : receiverReplacement, substMap, typeMap) {
+      Contract.Requires(substMap != null);
+      Contract.Requires(typeMap != null);
+    }
+    protected override List<BoundVar> CreateBoundVarSubstitutions(List<BoundVar> vars, bool forceSubstitutionOfBoundVars) {
+      var newBoundVars = vars.Count == 0 ? vars : new List<BoundVar>();
+      foreach (var bv in vars) {
+        var tt = Resolver.SubstType(bv.Type, typeMap);
+        var newBv = new BoundVar(bv.tok, "_'" + bv.Name, tt);
+        newBoundVars.Add(newBv);
+        // update substMap to reflect the new BoundVar substitutions
+        var ie = new IdentifierExpr(newBv.tok, newBv.Name);
+        ie.Var = newBv;  // resolve here
+        ie.Type = newBv.Type;  // resolve here
+        substMap.Add(bv, ie);
+      }
+      return newBoundVars;
+    }
+  }
+}
